+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 1.2.1
+
+- Set OCAMLFIND_CONF and PATH (or LD_LIBRARY_PATH on Unix) for
+  `ocaml`, `utop`, `utop-full`, `down` and `ocamlfind` shims
+- Remove PATH addition for fswatch from `dune` shim
+
 ## 1.0.2
 
 - On `*nix` use `with-dkml` binary rather than `with-dkml.exe`. Same for `dkml-fswatch` which is not